Code was executed through Tensorflow environment
Code is not fixed yet and can be seen that it's throwing several errors. It was executed through a newly-created environment 'Tensorflow' and had to download tensorflow and other packages since it's not letting me download it to the root environment.  Still working on it and it should be fixed by tomorrow.